functions/requirements.txt
stripe
firebase-admin
aksharamukha-transliterate
